HW-4/analyze-ads-cassandra/requirements.txt
pandas
python-dateutil
gdown
python-dotenv
rich
cassandra-driver